+++ lib/src/widgets/video_player_renderer.dart
@@ -60,10 +60,12 @@ class VideoPlayerRendererState extends State<VideoPlayerRenderer> {
   Widget build(BuildContext context) {
     final theme = OmniVideoPlayerTheme.of(context)!;
 
-    final aspectRatio = (widget.controller.rotationCorrection == 90 ||
-            widget.controller.rotationCorrection == 270)
-        ? widget.controller.size.height / widget.controller.size.width
-        : widget.controller.size.width / widget.controller.size.height;
+    final aspectRatio =
+        widget.options.playerUIVisibilityOptions.customAspectRatioNormal ??
+            ((widget.controller.rotationCorrection == 90 ||
+                    widget.controller.rotationCorrection == 270)
+                ? widget.controller.size.height / widget.controller.size.width
+                : widget.controller.size.width / widget.controller.size.height);
 
     return VideoOverlayControls(
       controller: widget.controller,
@@ -120,6 +122,7 @@ class VideoPlayerRendererState extends State<VideoPlayerRenderer> {
                     child: AdaptiveVideoPlayerDisplay(
                       controller: widget.controller,
                       isFullScreenDisplay: false,
+                      aspectRatio: aspectRatio,
                     ),
                   ),
                 ),